Redis Cache

+++ api/photos.js
@@ -1,5 +1,5 @@
 const express = require("express");
-// const Redis = require("ioredis");
+const { fetchFromCache, invalidateCache }  = require('./cacheHelper')
 const fs = require("fs");
 const exiftool = require("exiftool-vendored").exiftool;
 const multer = require("multer");
@@ -14,68 +14,41 @@ const {
 } = require("../db/models");
 const { Op, json } = require("sequelize");
 
-//const { Sequelize } = require("sequelize");
+const { Sequelize } = require("sequelize");
+// const { default: Redis } = require("ioredis");
 //create new redis instance
 // const redis = new Redis(process.env.REDIS_URL);
 
-//Root here is localhost:8000/api/photos
-//getting all photos
-// router.get("/", async (req, res, next) => {
-//     try {
-//         let allPhotos = null;
-//         const startTimeRedis = process.hrtime(); // Get the start time
-//         const cachedPosts = await redis.get("allPhotos");
-//         if(cachedPosts){
-//             allPhotos = JSON.parse(cachedPosts);
-//             const endTimeRedis = process.hrtime(startTimeRedis); // Get the end time for Redis retrieval
-//             const loadingTimeRedis = endTimeRedis[0] * 1000 + endTimeRedis[1] / 1e6; // Calculate the Redis loading time in milliseconds
-//             console.log("Redis Loading time:", loadingTimeRedis, "ms");
-//         }else{
-//             const startTimeDB = process.hrtime();
-//             allPhotos = await Photo.findAll({
-//                 //if you want to randomize the data, - also uncomment sequelize too
-//                 //order: [[Sequelize.literal('RAND()')]],
-//                 limit: 30,
-//                 //all the includes
-//                 include: { all: true, nested: true },
-//             });
-//             const endTimeDB = process.hrtime(startTimeDB); // Get the end time for direct database retrieval
-//             const loadingTimeDB = endTimeDB[0] * 1000 + endTimeDB[1] / 1e6; // Calculate the database loading time in milliseconds
-//             console.log("Database Loading time:", loadingTimeDB, "ms");
-//             if(allPhotos){
-//                 await redis.set("allPhotos", JSON.stringify(allPhotos));
-//             }
-//         }
-//         allPhotos
-//             ? res.status(200).json(allPhotos)
-//             : res.status(404).send("Photos Not Found");
-//     } catch (error) {
-//         next(error);
-//     }
-// });
+//Root here is localhost:8000/api/photo
 router.get("/", async (req, res, next) => {
   try {
-    const allphotos = await Photo.findAll({
-      //if you want to randomize the data, - also uncomment sequelize too
-      //order: [[Sequelize.literal('RAND()')]],
-      limit: 30,
-      //all the includes
-      include: { all: true, nested: true },
+    const allPhotos = await fetchFromCache('allPhotos', () => {
+      return Photo.findAll({
+        //if you want to randomize the data, - also uncomment sequelize too
+        //order: [[Sequelize.literal('RAND()')]],
+        limit: 30,
+        //all the includes
+        include: { all: true, nested: true },
+      });
     });
-    allphotos
-      ? res.status(200).json(allphotos)
+
+    allPhotos
+      ? res.status(200).json(allPhotos)
       : res.status(404).send("Photos Not Found");
   } catch (error) {
     next(error);
   }
 });
 
+
 router.get("/:id", async (req, res, next) => {
   const photoId = req.params.id;
   try {
-    const singlePhoto = await Photo.findByPk(photoId, {
-      //all the includes
-      include: { all: true, nested: true },
+    const singlePhoto = await fetchFromCache(`photo:${photoId}`, () => {
+      return Photo.findByPk(photoId, {
+        //all the includes
+        include: { all: true, nested: true },
+      });
     });
     singlePhoto
       ? res.status(200).json(singlePhoto)
@@ -85,15 +58,20 @@ router.get("/:id", async (req, res, next) => {
   }
 });
 
+
 router.get("/user/:id", async (req, res, next) => {
   const { id } = req.params;
   try {
-    const photos = await Photo.findAll({ where: { userId: id } });
+    const photos = await fetchFromCache(`user:${id}:photos`, () => {
+      return Photo.findAll({ where: { userId: id } });
+    });
     res.json(photos);
   } catch (error) {
     next(error);
   }
 });
+
+
 //add photo
 router.post("/addPhoto", async (req, res, next) => {
   try {
@@ -146,6 +124,9 @@ router.post("/addPhoto", async (req, res, next) => {
       locationId,
       cameraDetailId,
     });
+
+    await invalidateCache("allPhotos", `user:${userId}:photos`);
+
     createPhoto
       ? res.status(200).json(createPhoto)
       : res.status(400).send("Can't add photo");
@@ -214,6 +195,9 @@ router.put("/:id", async (req, res, next) => {
     const updatedPhoto = await Photo.update(updatePhoto, {
       where: { id: photoId },
     });
+
+    await invalidateCache("allPhotos");
+
     updatedPhoto
       ? res.status(200).json(updatedPhoto)
       : res.status(400).send("Photo not found");
@@ -225,9 +209,22 @@ router.put("/:id", async (req, res, next) => {
 router.delete("/:id", async (req, res, next) => {
   const photoId = req.params.id;
   try {
+
+    const photo = await Photo.findByPk(photoId);
+
+    if (!photo) {
+      return res.status(400).send("Photo not found");
+    }
+
+    const userId = photo.userId; 
+
     const deletedPhoto = await Photo.destroy({
       where: { id: photoId },
     });
+
+    await invalidateCache("allPhotos");
+    await invalidateCache(`user:${userId}:photos`)
+    // await invalidateCache(`user:${userId}:likes`);
 
     deletedPhoto
       ? res.status(200).send("Photo deleted")
@@ -254,7 +251,7 @@ router.get("/tag/:tagId", async (req, res, next) => {
 // Like photo
 router.post("/:photoId/like", async (req, res, next) => {
   try {
-    const userId = req.body.userId; // let's assume you send userId in the request body
+    const userId = req.body.userId; 
     const photoId = req.params.photoId;
     console.log("Request body:", req.body);
     console.log(`Attempting to like photo ${photoId} for user ${userId}`);
@@ -266,6 +263,7 @@ router.post("/:photoId/like", async (req, res, next) => {
     photo.likesCount += 1;
     await photo.save();
 
+    await invalidateCache(`user:${userId}:likes`)
     res.status(200).json(like);
   } catch (error) {
     //console.error(`Failed to like photo ${photoId} for user ${userId}`, error);
@@ -282,12 +280,17 @@ router.delete("/:photoId/unlike", async (req, res, next) => {
       `User with ID: ${userId} is unliking photo with ID: ${photoId}`
     );
     await Like.destroy({ where: { userId, photoId } });
-    res.status(200).send("Photo unliked");
 
     // Decrement the likesCount
     const photo = await Photo.findByPk(photoId);
     photo.likesCount -= 1;
     await photo.save();
+    
+    // await invalidateCache("allPhotos");
+    // await invalidateCache(`user:${userId}:photos`);
+    await invalidateCache(`user:${userId}:likes`);
+
+    res.status(200).send("Photo unliked")
     
   } catch (error) {
     next(error);
